coords: refactor transformations as module (#15895)
---------
