+++ docs/changes/coordinates/15895.api.rst
@@ -1,0 +1,3 @@
+The ``astropy.coordinates.transformations`` module has been refactored into a module.
+There should be no user-visible changes, but if you notice any, please open an
+Issue.
